Move "TextSetting" to new folder
