+++ src/components/settings/text/TextSetting.tsx
